Updated Asset Browser
Updated Asset Browser to look more like unity.
Added search functionality.
Added icons with fallback, to get icons, add a folder to the /bin/assets called icons, then add the png's with the appropriate names, look at the function LoadAssetIcons() on line 4475 in EditorMain.cpp for the names the icons need to be named.