Updated Editor
Updated multiple panels in the editor so make it so they don't run code if they are not active and visable, thus saving performance.
Also improved the resource panel, so you can select and see the name of the resource and added icons for it.
Improved the icons for both resource and asset panel so we have a fall back default icon, if the default icon can't be found, use text instead.